Code style fixes and SendGridAccessibilityService implementation
