+++ OutOfSchool/OutOfSchool.AuthCommon/Controllers/AccountController.cs
@@ -4,6 +4,7 @@ using Microsoft.Extensions.Options;
 using OutOfSchool.AuthCommon.Config;
 using OutOfSchool.AuthCommon.Extensions;
 using OutOfSchool.AuthCommon.ViewModels;
+using OutOfSchool.EmailSender.Services;
 using OutOfSchool.RazorTemplatesData.Models.Emails;
 
 namespace OutOfSchool.AuthCommon.Controllers;
