+++ OutOfSchool/OutOfSchool.AuthServer.Tests/Controllers/RegionAdminControllerTests .cs
@@ -20,10 +20,10 @@ using Microsoft.AspNetCore.Identity;
 using OutOfSchool.RazorTemplatesData.Services;
 using Microsoft.Extensions.Localization;
 using OutOfSchool.AuthCommon;
-using OutOfSchool.EmailSender;
 using System.Linq;
 using Microsoft.AspNetCore.Identity.EntityFrameworkCore;
 using Microsoft.EntityFrameworkCore.Diagnostics;
+using OutOfSchool.EmailSender.Services;
 
 namespace OutOfSchool.AuthServer.Tests.Controllers;
