+++ OutOfSchool/OutOfSchool.EmailSender/EmailSenderStringConstants.cs
@@ -1,0 +1,10 @@
+﻿namespace OutOfSchool.EmailSender;
+
+public static class EmailSenderStringConstants
+{
+    public const string Email = "Email";
+    public const string Subject = "Subject";
+    public const string HtmlContent = "HtmlContent";
+    public const string PlainContent = "PlainContent";
+    public const string ExpirationTime = "ExpirationTime";
+}
