+++ OutOfSchool/OutOfSchool.EmailSender/Services/DevEmailSender.cs
@@ -2,7 +2,7 @@ using System;
 using System.Threading.Tasks;
 using Microsoft.Extensions.Logging;
 
-namespace OutOfSchool.EmailSender;
+namespace OutOfSchool.EmailSender.Services;
 
 public class DevEmailSender : IEmailSenderService
 {
@@ -12,18 +12,18 @@ public class DevEmailSender : IEmailSenderService
     {
         this.logger = logger;
     }
-        
+
     public Task SendAsync(string email, string subject, (string html, string plain) content, DateTime? expirationTime)
     {
         // This code runs only in dev for testing purposes
         // so can ignore "not log user-controlled data".
-        #pragma warning disable S5145
+#pragma warning disable S5145
         logger.LogDebug(
             "Sending mail to {Email} with subject '{Subject}' and content: {Content}",
             email,
             subject,
             content.html);
-        #pragma warning restore S5145
+#pragma warning restore S5145
         return Task.CompletedTask;
     }
 }
