+++ OutOfSchool/OutOfSchool.EmailSender/Services/ISendGridAccessibilityService.cs
@@ -1,0 +1,9 @@
+﻿using System;
+
+namespace OutOfSchool.EmailSender.Services;
+
+public interface ISendGridAccessibilityService
+{
+    bool IsSendGridAccessible { get; set; }
+    DateTimeOffset? GetNextStartDate();
+}
